BaseSpriteInterface, TextEffectTeletype, updated Animation, AnimationByFile, Text moved to Sprites/TextClass, animations moved to Sprites/AnimationClass, metadata for windows, fixed mouse_pos (int Letterboxing_w now), updated BaseSprite::is_collided_with_sprite

+++ SDL3Game/main.cpp
@@ -91,7 +91,7 @@ SDL_AppResult handleSceneEvents(EventType event) {
     return SDL_APP_CONTINUE;
 }
 
-float Letterboxing_w, Letterboxing_h;
+int Letterboxing_w, Letterboxing_h;
 SDL_AppResult SDL_AppEvent(void* appstate, SDL_Event* event)
 {
     switch (event->type)
